Add basic document parser and validator.
Including UTs and sample good and bad files.

Realised there was a problem with my previous schema (or my
interpretation of it), to do with elementFormDefault. Nested elements
were unqualified in the schema, but I was using xmlns in my document
which meant I was qualifying them. Leads to a truly bizarre error
message of the form "Invalid content was found starting with element
'foo'. One of '{foo}' is expected."